edit and adds are modified

+++ tommystore/src/com/tommystore/controller/CategoryController.java
@@ -4,6 +4,8 @@ import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -12,14 +14,26 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.tommystore.domain.Category;
+import com.tommystore.exceptions.ObjectDeletionException;
 import com.tommystore.service.CategoryService;
+import com.tommystore.service.ProductService;
 
 @Controller
 @RequestMapping(value = "/admin")
+@PropertySource("/WEB-INF/properties")
 public class CategoryController {
 
 	@Autowired
 	private CategoryService categoryService;
+	
+	@Autowired
+	private ProductService productService;
+	
+	@Value("${invalid.categoryNameUsed}")
+	String nameUsedErrorMessage;
+	
+	@Value("${invalid.categoryDeletion}")
+	String deletionErrorMessage;
 	
     @RequestMapping(value = "/category-list-view", method = RequestMethod.GET)
     public String categoryView(Model model) {
@@ -40,6 +54,11 @@ public class CategoryController {
             return "admin-dashboard-add-category";
         }	
         
+        if(categoryService.isCategoryExistByName(category.getName())) {
+        	model.addAttribute("message", nameUsedErrorMessage);
+            return "admin-dashboard-add-category";
+        }
+        
         categoryService.saveCategory(category);
 		return "redirect:dashboard";
     }
@@ -55,18 +74,35 @@ public class CategoryController {
     public String editCategory(@Valid Category categoryToEdit, BindingResult result, Model model) {
     	
         if (result.hasErrors()) {
-            return "admin-dashboard-edit-product";
-        }	
+        	Category category = categoryService.findCategoryById(categoryToEdit.getId());
+        	model.addAttribute("category", category);
+        	return "admin-dashboard-edit-category";
+        }
+        
+        if(!categoryService.isValidToEditByIdAndName(categoryToEdit.getId(), categoryToEdit.getName())) {
+        	Category category = categoryService.findCategoryById(categoryToEdit.getId());
+        	model.addAttribute("message", nameUsedErrorMessage);
+        	model.addAttribute("category", category);
+        	return "admin-dashboard-edit-category";
+        }
+        
         Category category = categoryService.findCategoryById(categoryToEdit.getId());
         category.setName(categoryToEdit.getName());
         categoryService.saveCategory(category);
+        
 		return "redirect:category-list-view";
     }
     
     @RequestMapping(value = "/delete-category", method = RequestMethod.GET)
-    public String deleteCategory(Model model, @RequestParam("id") Integer id) {
+    public String deleteCategory(Model model, @RequestParam("id") Integer id) throws ObjectDeletionException {
+
+		if(productService.findProductByCategory(id).size() != 0) {
+	    	model.addAttribute("categoryList", categoryService.getCategoryList());
+	    	model.addAttribute("message", deletionErrorMessage);
+			return "admin-dashboard-categorylist";
+		}
     	categoryService.deleteCategory(id);
+    	
 		return "redirect:category-list-view";
     }
-	
 }
